+++ test/smokes/cppcheck/expectations.rb
@@ -413,3 +413,38 @@ s.add_test(
   "unexpected_error",
   type: "failure", message: "cppcheck: Unknown language 'foo' enforced.", analyzer: { name: "Cppcheck", version: :_ }
 )
+
+s.add_test(
+  "addon",
+  type: "success",
+  issues: [
+    {
+      id: "misra-c2012-14.4",
+      path: "bad.c",
+      location: { start_line: 8 },
+      message: "misra violation (use --rule-texts=<file> to get proper output)",
+      links: [],
+      object: { severity: "style", verbose: nil, inconclusive: false, cwe: nil, location_info: nil },
+      git_blame_info: nil
+    },
+    {
+      id: "misra-c2012-21.1",
+      path: "bad.h",
+      location: { start_line: 2 },
+      message: "misra violation (use --rule-texts=<file> to get proper output)",
+      links: [],
+      object: { severity: "style", verbose: nil, inconclusive: false, cwe: nil, location_info: nil },
+      git_blame_info: nil
+    },
+    {
+      id: "y2038-type-bits-undef",
+      path: "bad.h",
+      location: { start_line: 2 },
+      message: "_USE_TIME_BITS64 is defined but _TIME_BITS was not",
+      links: [],
+      object: { severity: "warning", verbose: nil, inconclusive: false, cwe: nil, location_info: nil },
+      git_blame_info: nil
+    }
+  ],
+  analyzer: { name: "Cppcheck", version: :_ }
+)